Revert "rebase: introduce a shortcut for --reschedule-failed-exec"
When contributing the patch series, the cover letter tried to convey
clearly that the patch introducing the shortcut -y was included only to
show that it is possible, with a slight bias against it.

During the review, there were a couple reviewers who agreed with this
sentiment, and the author was happy that this patch was not needed and
concurred that it should be dropped. See e.g. Stefan Beller's reply:
<CAGZ79kZL5CRqCDRb6B-EedUm8Z_i4JuSF2=UtwwdRXMitrrOBw@mail.gmail.com>

However, it slipped by the original patch author (yours truly) that the
patch *was* included when the branch made it to `next` and then when it
made it to `master`.

So let's back out that patch before it even slips into an official
release (in which case we would even have to support this unwanted
flag).

This reverts commit 81ef8ee75d5f348d3c71ff633d13d302124e1a5e.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>